external/biomed-multi-view/requirements.txt
setuptools==69.5.1
torch==2.1.0
torchvision==0.16.0
fuse-med-ml==0.3.0
jsonargparse[signatures]==4.33.2
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
rdkit-pypi==2022.9.5
regex==2024.9.11
biopython==1.84
datasets==2.12.0
pyarrow==12.0.0
openbabel-wheel==3.1.1.20
transformers==4.45.1
imageio==2.35.1
loralib==0.1.2
mplcursors==0.5.3
xgboost==2.0.3
joblib==1.3.2
x-transformers==1.27.19
lmdb==1.4.1
timm==0.9.16
litdata==0.2.6
torch_geometric==2.3.1
click==8.1.7
huggingface_hub==0.25.1
pyg_lib
torch_scatter
torch_cluster
torch_spline_conv
-f https://data.pyg.org/whl/torch-2.1.0+cu121.html
pytorch-fast-transformers==0.4.0